[TASK] Fix phpstan checkFunctionArgumentTypes errors in ext:backend Form
This patch fixes incompatible type usage in function arguments
and is preparatory work for introducing native type hints and
strict mode in all core files.

Releases: master, 10.4
Resolves: #92579
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66194
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>

+++ typo3/sysext/backend/Classes/Form/FormDataProvider/DatabaseRowDateTimeFields.php
@@ -44,7 +44,7 @@ class DatabaseRowDateTimeFields implements FormDataProviderInterface
                     // Create an ISO-8601 date from current field data; the database always contains UTC
                     // The field value is something like "2016-01-01" or "2016-01-01 10:11:12", so appending "UTC"
                     // makes date() treat it as a UTC date (which is what we store in the database).
-                    $result['databaseRow'][$column] = date('c', strtotime($result['databaseRow'][$column] . ' UTC'));
+                    $result['databaseRow'][$column] = date('c', (int)strtotime($result['databaseRow'][$column] . ' UTC'));
                 } else {
                     $result['databaseRow'][$column] = null;
                 }
